augment: keep background noise in float32; finetune: add --num-workers
np.random.normal returns float64, so the add and the clip in random_background both ran at double width. Measured peak for a 256x256x3 draw: 4.50 MiB before, 2.25 MiB after.

Casting the draw rather than switching to default_rng was deliberate. Measured, the generator switch reaches the same 2.25 MiB and buys nothing further, while changing the random stream -- so the zero-risk variant is also the equal-value one. The handoff's 4.69 -> 1.13 MiB figure was not reproducible; 2.25 MiB is the floor for this expression.

finetune.py gains the --num-workers that train.py already had, so concurrent arms can be memory-bounded the same way.

+++ src/training/augment.py
@@ -32,7 +32,9 @@ def random_background(size, rng):
     else:
         small = np.random.randint(0, 256, (max(2, h // 32), max(2, w // 32), 3), dtype=np.uint8)
         bg = np.array(Image.fromarray(small).resize((w, h), Image.BICUBIC))
-    noise = np.random.normal(0, 8, bg.shape)
+    # normal() yields float64, which doubles the peak through the add and the clip.
+    # Casting rather than switching generator keeps the draw sequence identical.
+    noise = np.random.normal(0, 8, bg.shape).astype(np.float32)
     return np.clip(bg.astype(np.float32) + noise, 0, 255).astype(np.uint8)
